third_party/WebKit/LayoutTests/svg/animations/pausing-use-counter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg id="svgRoot" width="400" height="400">
</svg>
<script>
'use strict';

// From UseCounter.h
var SVGSMILPausing = 1506;
var SVGSMILCurrentTime = 1507;

test(() => {
    var svgRoot = document.getElementById('svgRoot');
    assert_false(internals.isUseCounted(document, SVGSMILCurrentTime));
    svgRoot.getCurrentTime();
    assert_true(internals.isUseCounted(document, SVGSMILCurrentTime));

    assert_false(internals.isUseCounted(document, SVGSMILPausing));
    svgRoot.animationsPaused();
    assert_true(internals.isUseCounted(document, SVGSMILPausing));
}, 'animationsPaused is use counted');

</script>
