third_party/WebKit/LayoutTests/platform/win/svg/animations/pausing-use-counter-expected.txt

This is a testharness.js-based test.
PASS animationsPaused is use counted 
Harness: the test ran to completion.
